chore: rename tailwind module to django_tailwind
The current name clashes with tailwind module.
