+++ src/django_tailwind/app_template/cookiecutter.json
