+++ src/django_tailwind/app_template/{{cookiecutter.app_name}}/apps.py
