+++ src/django_tailwind/app_template/{{cookiecutter.app_name}}/static_src/package.json
