+++ src/django_tailwind/app_template/{{cookiecutter.app_name}}/templates/base.html
